requirements.txt
# Notion-Obsidian同期のためのコア依存関係
notion-client==2.2.1
PyYAML==6.0.1
python-dotenv==1.0.0

# 開発とテストの依存関係
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# 拡張機能のためのオプション依存関係
requests==2.31.0